+++ package.json
@@ -2,18 +2,21 @@
   "name": "ff-nodejs-server-sdk",
   "version": "0.0.1",
   "description": "Feature flags SDK for NodeJS environments",
-  "main": "dist/index.js",
+  "main": "dist/sdk.cjs",
+  "module": "dist/sdk.mjs",
   "types": "dist/index.d.ts",
   "engines": {
     "node": ">= 12"
   },
   "devDependencies": {
+    "@openapitools/openapi-generator-cli": "^2.3.10",
     "@types/eventsource": "^1.1.6",
     "@types/jest": "~26.0.23",
     "@types/keyv": "^3.1.3",
     "@types/node": "^14.17.11",
     "@typescript-eslint/eslint-plugin": "~4.28.2",
     "@typescript-eslint/parser": "~4.28.2",
+    "esbuild": "^0.13.2",
     "eslint": "~7.30.0",
     "eslint-config-prettier": "~8.3.0",
     "eslint-plugin-jest": "~24.3.6",
@@ -28,7 +31,9 @@
     "clean": "rimraf coverage dist tmp",
     "versioning": "node -p \"'export const VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/version.ts",
     "prebuild": "npm run clean; npm run versioning; npm run lint",
-    "build": "tsc -p tsconfig.release.json",
+    "build:esm": "esbuild ./src/index.ts --tsconfig=tsconfig.modules.json --minify --bundle --target=node12.22.5 --platform=node --format=esm --external:fs --outfile=./dist/sdk.mjs",
+    "build:cjs": "esbuild ./src/index.ts --tsconfig=tsconfig.release.json --minify --bundle --target=node12.22.5 --platform=node --format=cjs --external:fs --outfile=./dist/sdk.cjs",
+    "build": "npm run build:esm; npm run build:cjs",
     "build:watch": "tsc -w -p tsconfig.release.json",
     "lint": "eslint ./src --ext .ts,.tsx",
     "test": "jest --coverage",
@@ -38,7 +43,6 @@
   "author": "Enver Bisevac <enver.bisevac@harness.io>",
   "license": "Apache-2.0",
   "dependencies": {
-    "@openapitools/openapi-generator-cli": "^2.3.10",
     "eventsource": "^1.1.0",
     "jwt-decode": "^3.1.2",
     "keyv": "^4.0.3",
