+++ tsconfig.modules.json
@@ -1,0 +1,12 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "target": "ES6",
+        "module": "ES6",
+        "outDir": "dist",
+        "removeComments": true
+    },
+    "include": [
+        "src/**/*.ts"
+    ]
+}
